Move fileContainsString() to testutils.go
Multiple tests will use this function, so it makes
sense to put it in a separate file
