+++ pkg/environment/templateEnvironment_test.go
@@ -1,10 +1,7 @@
 package environment
 
 import (
-	"fmt"
-	"io/ioutil"
 	"os"
-	"strings"
 	"testing"
 )
 
@@ -15,17 +12,6 @@ func cleanUpNamespacesFolder(namespace string) {
 	os.Remove(namespaceFolder)
 	os.Remove(namespaceBaseFolder)
 	os.Remove("namespaces")
-}
-
-func fileContainsString(t *testing.T, filename string, searchString string) {
-	contents, err := ioutil.ReadFile(filename)
-	if err != nil {
-		t.Errorf("Unexpected error: %s", err)
-	}
-
-	if !(strings.Contains(string(contents), searchString)) {
-		t.Errorf(fmt.Sprintf("Didn't find %s in contents of %s", searchString, filename))
-	}
 }
 
 func TestCreateNamespace(t *testing.T) {
